requirements.txt
f5-tts
streamlit
python-docx
chardet
torch
